+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+### 2.1.0 (2022-04-23)
+
+##### New Features
+
+* **self-banking:**  admin-99 users can now bank for services even if they have previously not banked ([61514864](https://github.com/firstlovecenter/fl-admin-portal/commit/615148649a23968e929a867fc3bacedee6ae83eb))
+
+##### Bug Fixes
+
+* **self-banking:**
+  *  fixed object structure with confirm Service state ([a57dff2b](https://github.com/firstlovecenter/fl-admin-portal/commit/a57dff2b40196d8648aa05c56ce257ec61b5c7aa))
+  *  console logging response from confirm payment mutation ([afe12b23](https://github.com/firstlovecenter/fl-admin-portal/commit/afe12b2340574c6f092ac7dac906165e14157737))
+
+#### 2.0.9 (2022-04-22)
+
+##### New Features
+
+* **self-banking:**
+  *  you can no longer bank if you have a previous pending transaction ([1afb879a](https://github.com/firstlovecenter/fl-admin-portal/commit/1afb879a6d0a3c6f005b787e5561aa16d4ea9db8))
+  *  improved UX by handling for error states ([b524f86f](https://github.com/firstlovecenter/fl-admin-portal/commit/b524f86f68731220204c4f3e267f042443bcece8))
+
 #### 2.0.8 (2022-04-22)
 
 ##### Bug Fixes